+++ ReleaseNotes.md
@@ -1,4 +1,9 @@
 
+## New Features
+
+- [#298] WebUI - Add configuration support to modify the host and port both the API and WebUI bind to
+
 ## Fixes
 
 - [#292] WebUI - Fixing `Failed to get settings from db` error message on fresh install
+- [#296] Docs site fixes